+++ frontend/apps/redux-counter/src/store/index.ts
@@ -1,32 +1,31 @@
-// Importing the Redux toolkit configuration and the reducers.
 import { configureStore, Store } from '@reduxjs/toolkit';
-import counterReducer from './counterSlice';
-import authenticationSlice from './authenticationSlice';
+import counter from './slices/counter';
+import authentication from './slices/authentication';
+
 /**
  * Creates and configures the Redux store.
  * The store is created with a root reducer consisting of `unitReducer` and `statusSlice`.
- * @type {Store}
+ * @NOTE: The configureStore function is designed to infer types automatically, so you typically don't need to manually specify types with Store
  */
-
-// Configure the store with the counter slice reducer
 export const store: Store = configureStore({
     reducer: {
-        counter: counterReducer,
-        authentication: authenticationSlice,
+        counter,
+        authentication,
     },
 });
 
 /**
  * Infers and exports the `RootState` type from the store's state.
  * The `RootState` type represents the complete state of the Redux store.
- * @type {State}
  */
-
 export type RootState = ReturnType<typeof store.getState>;
 
 /**
  * Infers and exports the `AppDispatch` type from the store's dispatch function.
  * The `AppDispatch` type represents the dispatch function to interact with the Redux store.
- * @type {Dispatch}
  */
 export type AppDispatch = typeof store.dispatch;
+export type AppStore = typeof store;
+
+export const isAuthenticatedSelector = (state: RootState) => state.authentication;
+export const counterSelector = (state: RootState) => state.counter;